requirements.txt
asgiref==3.8.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==2.0.0
Flask-Cors==4.0.1
Frozen-Flask==1.0.2
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.0.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
pydrive2==1.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==2.0.0
